chore: mv tag-input drag interface
