+++ src/tag-input/hooks/useDragSorter.ts
@@ -1,5 +1,5 @@
 import { ref, reactive } from 'vue';
-import { DragSortProps, DragSortInnerProps } from '../type';
+import { DragSortProps, DragSortInnerProps } from '../interface';
 
 export default function useDragSorter<T>(props: DragSortProps<T>): DragSortInnerProps {
   const { sortOnDraggable, onDragSort, onDragOverCheck } = props;
