+++ src/tag-input/type.ts
@@ -3,7 +3,6 @@
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
-import { Ref } from 'vue';
 import { InputProps } from '../input';
 import { InputValue } from '../input';
 import { TagProps } from '../tag';
@@ -200,34 +199,3 @@ export interface TagInputRemoveContext {
 }
 
 export type TagInputRemoveTrigger = 'tag-remove' | 'backspace';
-
-interface DragSortContext<T> {
-  currentIndex: number;
-  current: T;
-  targetIndex: number;
-  target: T;
-}
-
-export interface DragSortProps<T> {
-  sortOnDraggable: boolean;
-  onDragSort?: (context: DragSortContext<T>) => void;
-  onDragOverCheck?: {
-    x?: boolean;
-    targetClassNameRegExp?: RegExp;
-  };
-}
-
-type DragFnType = (e?: DragEvent, index?: number, record?: any) => void;
-interface DragSortInnerData {
-  dragging?: boolean;
-  onDragStart?: DragFnType;
-  onDragOver?: DragFnType;
-  onDrop?: DragFnType;
-  onDragEnd?: DragFnType;
-}
-
-export interface DragSortInnerProps extends DragSortInnerData {
-  getDragProps?: (index?: number, record?: any, tagRef?: Ref<HTMLElement>) => DragSortInnerData;
-}
-
-export interface TagInputProps extends TdTagInputProps, DragSortInnerProps {}
